+++ ui/src/main/java/io/snabble/sdk/ui/cart/shoppingcart/cartdiscount/CartDiscount.kt
@@ -3,8 +3,10 @@ package io.snabble.sdk.ui.cart.shoppingcart.cartdiscount
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.material3.Surface
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -13,7 +15,10 @@ import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.unit.dp
+import io.snabble.sdk.checkout.LineItem
+import io.snabble.sdk.shoppingcart.ShoppingCart
 import io.snabble.sdk.ui.cart.shoppingcart.cartdiscount.model.CartDiscountItem
 import io.snabble.sdk.ui.cart.shoppingcart.cartdiscount.widget.DiscountDescription
 
@@ -38,6 +43,21 @@ internal fun CartDiscount(
             title = stringResource(item.title),
             description = item.name,
             discount = item.discount
+        )
+    }
+}
+
+@PreviewLightDark
+@Composable
+private fun Preview() {
+    Surface {
+        CartDiscount(
+            modifier = Modifier.fillMaxWidth(),
+            item = CartDiscountItem(
+                item = ShoppingCart.Item(ShoppingCart(), LineItem(id = "", amount = 1)),
+                discount = "7.00",
+                name = "SUPER DUPER RABATT"
+            )
         )
     }
 }